Move "spew" code into its own file.
